pyproject.toml
[project]
name = "expense-tracker-api-flask"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.2",
    "psycopg-pool>=3.2.6",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.9",
    "pyjwt>=2.10.1",
]

[dependency-groups]
dev = [
    "faker>=37.8.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-benchmark>=5.1.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.13.3",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]


[tool.ruff]
line-length = 88
target-version = "py312"
exclude = [
  "build/",
  "dist/",
  ".git/",
  "__pycache__",
  "migrations/",
  ".venv/",
  ".ruff_cache"
]


#[tool.ruff.lint]
#select = ["ALL"]


[tool.ruff.lint]
select = [
    "E", "W", "F", "I", "N", "B", "S", "UP",
    "C90", "C4", "SIM", "RET", "TRY", "PL",
    "LOG", "Q", "RUF"
]
ignore = [
    "E501",
    "D1",
    "S311",
    "TRY003"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "S101",   # allow assert in tests
    "S105",   # ignore "hardcoded password" in test strings
]

[tool.mypy]
python_version = 3.12
warn_unused_configs = true
warn_return_any = true
disallow_untyped_defs = true
check_untyped_defs = true
strict_equality = true
ignore_missing_imports = true
exclude = "tests/|migrations/|.venv/"